BF: move reused _basic_publishmeta into dandischema/datacite/__init__.py + fix typing issue
